🚚 Rename: 마이페이지 디렉토리 정리 (#110)
